+++ CHANGELOG.md
@@ -58,3 +58,23 @@ Updated gitignore file
  Added coverage test
 
 ### Fixed
+
+## [1.0.13] - 2022-02-01
+ 
+### Added
+
+ Added unit tests , integration tests 
+
+ Add inquirer support to allow interactive use (commandline arguments still supported)
+ 
+ Added coverage test
+
+### Fixed
+
+## [2.0.0] - 2022-02-01
+ 
+### Added
+
+Added capability to replace content using Regex
+
+### Fixed